# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Swagger Gentcmen API documentation",
        "contact": {
            "email": "nikitosnov@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/admin/product-categories": {
            "post": {
                "tags": [
                    "Product Categories"
                ],
                "summary": "Create product category",
                "operationId": "Create product category",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProductCategoryVirtualBody"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product status is created successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product Categories"
                ],
                "summary": "Delete some product category items",
                "operationId": "Delete some product category items",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DestroySomeItemsVirtualBody"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Product category items deleted successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/product-categories/{category}": {
            "put": {
                "tags": [
                    "Product Categories"
                ],
                "summary": "Update product category by id",
                "operationId": "Update product category",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Product category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProductCategoryVirtualBody"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product status is updated successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product Categories"
                ],
                "summary": "Delete product category by id",
                "operationId": "Delete product category by id",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Product category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Product category deleted successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/client/product-category-options/{productCategoryOption}": {
            "get": {
                "tags": [
                    "Product Category Options"
                ],
                "summary": "Fetch product category option by id",
                "operationId": "Fetch product category option by id",
                "parameters": [
                    {
                        "name": "productCategoryOption",
                        "in": "path",
                        "description": "Option id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product category option fetched successfully"
                    },
                    "404": {
                        "description": "Product category option not found"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                }
            }
        },
        "/api/v1/admin/product-categories/{categoryId}/options": {
            "post": {
                "tags": [
                    "Product Categories"
                ],
                "summary": "Create product category filter option",
                "operationId": "Create product",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "Product category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProductCategoryOptionVirtualBody"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product category filter option is created successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/product-category-options/{option}": {
            "put": {
                "tags": [
                    "Product Category Options"
                ],
                "summary": "Update product category filter option",
                "operationId": "Update product category option",
                "parameters": [
                    {
                        "name": "option",
                        "in": "path",
                        "description": "Product category option id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProductCategoryOptionVirtualBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product category option is updated successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    },
                    "404": {
                        "description": "Product category option not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product Category Options"
                ],
                "summary": "Destroy product category option",
                "operationId": "Destroy product category option",
                "parameters": [
                    {
                        "name": "option",
                        "in": "path",
                        "description": "Product category option id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Product category option is deleted successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    },
                    "404": {
                        "description": "Product category option not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/admin/products": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create product",
                "operationId": "Create product",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProductVirtualBody"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product is created successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    },
                    "404": {
                        "description": "Product status not found with provided id"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete some product items",
                "operationId": "Delete some product items",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DestroySomeItemsVirtualBody"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Products were deleted successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/admin/products/{product}": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update product",
                "operationId": "Update product",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductVirtualBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product is updated successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete product by id",
                "operationId": "Delete product by id",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Product deleted successfully"
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateProductCategoryOptionVirtualBody": {
                "title": "Create product review example",
                "description": "Create product review body data",
                "required": [
                    "option_name"
                ],
                "properties": {
                    "option_name": {
                        "title": "option_name",
                        "description": "Option name",
                        "type": "string",
                        "example": "some option name"
                    },
                    "option_values": {
                        "title": "option_values",
                        "description": "Option values",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            {
                                "value": 20
                            }
                        ]
                    }
                },
                "type": "object"
            },
            "CreateProductCategoryVirtualBody": {
                "title": "Create product category example",
                "description": "Create product category body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the new product category",
                        "type": "string",
                        "example": "belts"
                    }
                },
                "type": "object"
            },
...
            "CreateProductVirtualBody": {
                "title": "Create product example",
                "description": "Create product body data",
                "required": [
                    "name",
                    "cost",
                    "amount",
                    "description",
                    "content",
                    "status_id",
                    "category_ids[]",
                    "meta_title",
                    "meta_description",
                    "meta_keywords",
                    "banner_image"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the new product",
                        "type": "string",
                        "example": "belt"
                    },
                    "cost": {
                        "title": "cost",
                        "description": "Cost of the new product",
                        "type": "integer",
                        "example": "200"
                    },
                    "amount": {
                        "title": "amount",
                        "description": "Amount of the new product",
                        "type": "integer",
                        "example": "12"
                    },
                    "description": {
                        "title": "description",
                        "description": "Description of the new product",
                        "type": "string",
                        "example": "some description"
                    },
                    "content": {
                        "title": "content",
                        "description": "Content of the new product",
                        "type": "object",
                        "example": {
                            "width": 20,
                            "strap_length": 20
                        }
                    },
                    "images_content": {
                        "title": "images_content",
                        "description": "images_content of the new product",
                        "type": "json",
                        "example": {
                            "": 40,
                            "color": "blue",
                            "length": "100",
                            "gender": "male",
                            "belt_type": "classic",
                            "clasp_type": "classic"
                        }
                    },
                    "status_id": {
                        "title": "status_ids",
                        "description": "Identifiers of statuses for the new product",
                        "type": "integer",
                        "example": 1
                    },
                    "category_ids[]": {
                        "title": "category_ids",
                        "description": "Identifiers of categories for the new product",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1
                        ]
                    },
                    "image[]": {
                        "title": "image",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "binary"
                        }
                    },
                    "meta_title": {
                        "title": "Meta title",
                        "type": "string",
                        "example": "meta title"
                    },
                    "meta_description": {
                        "title": "Meta description",
                        "type": "string",
                        "example": "meta description"
                    },
                    "meta_keywords": {
                        "title": "Meta keywords",
                        "type": "string",
                        "example": "meta keywords"
                    },
                    "banner_image": {
                        "title": "Banner image",
                        "type": "string",
                        "example": "banner image"
                    },
                    "published": {
                        "title": "published",
                        "description": "published or not banner",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "DestroySomeItemsVirtualBody": {
                "title": "Destroy some items example",
                "description": "Destroy some items body data",
                "required": [
                    "ids"
                ],
                "properties": {
                    "ids": {
                        "title": "ids",
                        "description": "Identifiers that should be deleted",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2,
                            3
                        ]
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}